Start-seDriver #101
There is still a lot of cleanup to do in the different browsers internal functions and reorganization so they look similar (same order of things). LogLevel need to be implemented in other browsers or Throw something if not implemented.

Won't be merged until every other items on V4.0 project are completed and this is CI validated + very well documented. (ETA still Sept. 04)